Add files via upload
This is a desktop application for beginners developed on Visual Studio. The term “HCJ live” resembles- ‘Html , CSS and Java Script’ live editor. This application allows you to code collaboratively in real time, much like live multiplayer editor that runs on different window panels.nn real-time, much like a live multi-player editor that runs on different window panels. 
 
This application works on Html, CSS and Java Script which are considered basic fundamental coding structures for the beginners to learn and which can provide convenient access of files, data and their storage paths.                                                    
*As this application’s development platform is Visual studio, (it is designed in C#). And if a beginner is new to the basic coding world and have hesitations while performing and creating programs : Hence this application is all what a beginner needs. 
 
And here is a rough snippet working of how this application operates- 
As its an application which is developed for the rookie students who are stepping in the all new world of code and different computing languages, its meant to make all their fuss go away and provide a convenient and swift way of coding.
As this application primarily operates on HTML, CSS and Java Script,( Hence working will be better explained by an example – if the user is scripting basic html code and do not have adequate knowledge about the tags used in it , he/she can simply insert tags by clicking the HTML button flashing on the top of the application window and further even if the user does not know about the CSS linking , this application will also do that with just click of a button.
We came up to the idea of making this application because early in our days we also has faced these challenges too and so we try and implement a reinforcement learning algorithm these leads to increase in our problem solving skills, and designed this by seeing our past experiences. After searching the internet we reached the understanding that solving these standard raising concerned issues towards the better understanding and learning never took as seriously as it should be done and was underrated, there should be no complications for development of a student.